Circular dependency resolution (#803)
* setup util base matrix class

* ran format

* rearranged devices a bit

* ran format

* reworked circular issue

* resolved test and ran format

* ran format

* added fixture

* fixed import

* removed factored matrix functionality from shared util

* fixed import

* added new matrix specific function
