+++ transformer_lens/FactoredMatrix.py
@@ -12,7 +12,7 @@ from typing import List, Tuple, Union, overload
 import torch
 from jaxtyping import Float
 
-import transformer_lens.utils as utils
+import transformer_lens.utilities.tensors as tensor_utils
 
 
 class FactoredMatrix:
@@ -169,7 +169,7 @@ class FactoredMatrix:
         """
         Ua, Sa, Vha = torch.svd(self.A)
         Ub, Sb, Vhb = torch.svd(self.B)
-        middle = Sa[..., :, None] * utils.transpose(Vha) @ Ub * Sb[..., None, :]
+        middle = Sa[..., :, None] * tensor_utils.transpose(Vha) @ Ub * Sb[..., None, :]
         Um, Sm, Vhm = torch.svd(middle)
         U = Ua @ Um
         Vh = Vhb @ Vhm
@@ -239,11 +239,11 @@ class FactoredMatrix:
         """
         return FactoredMatrix(
             self.U * self.S.sqrt()[..., None, :],
-            self.S.sqrt()[..., :, None] * utils.transpose(self.Vh),
+            self.S.sqrt()[..., :, None] * tensor_utils.transpose(self.Vh),
         )
 
     def get_corner(self, k=3):
-        return utils.get_corner(self.A[..., :k, :] @ self.B[..., :, :k], k)
+        return tensor_utils.get_corner(self.A[..., :k, :] @ self.B[..., :, :k])
 
     @property
     def ndim(self) -> int:
@@ -253,7 +253,7 @@ class FactoredMatrix:
         """
         Collapses the left side of the factorization by removing the orthogonal factor (given by self.U). Returns a (..., mdim, rdim) tensor
         """
-        return self.S[..., :, None] * utils.transpose(self.Vh)
+        return self.S[..., :, None] * tensor_utils.transpose(self.Vh)
 
     def collapse_r(self) -> Float[torch.Tensor, "*leading_dims ldim mdim"]:
         """
