+++ transformer_lens/utilities/devices.py
@@ -45,6 +45,18 @@ def get_device_for_block_index(
     return torch.device(device.type, device_index)
 
 
+def get_device():
+    if torch.cuda.is_available():
+        return torch.device("cuda")
+    if torch.backends.mps.is_available() and torch.backends.mps.is_built():
+        # Parse the PyTorch version to check if it's below version 2.0
+        major_version = int(torch.__version__.split(".")[0])
+        if major_version >= 2:
+            return torch.device("mps")
+
+    return torch.device("cpu")
+
+
 def move_to_and_update_config(
     model: Union[
         "transformer_lens.HookedTransformer",
